+++ Tasks/Content/JSONSchema.ps1
@@ -58,12 +58,6 @@
     ```
 #>
 Add-BuildTask -Name JSONSchema -Jobs {
-    # Scope can be lost when running Plumber on Plumber multiple times
-    if (-not (Get-Command Get-PlumberTaskFile -ErrorAction SilentlyContinue)) {
-        . (Join-Path $script:PlumberConfig.ModuleRoot 'Private/Test-PlumberTaskPathExcluded.ps1')
-        . (Join-Path $script:PlumberConfig.ModuleRoot 'Private/Get-PlumberTaskFile.ps1')
-    }
-
     if (-not (Get-Command Test-Json -ErrorAction SilentlyContinue)) {
         Write-Error 'Test-Json is required for JSON schema validation'
         return